requirements.txt
boto3>=1.24.86
huggingface-hub>=0.10.0
hydra-core>=1.2.0
joblib>=1.2.0
librosa>=0.9.2
mlflow>=1.29.0
numpy>=1.23.3
pesq==0.0.4
protobuf>=3.19.6
pystoi==0.3.3
pytest-lazy-fixture>=0.6.3
pytorch-lightning>=1.7.7
scikit-learn>=1.1.2
scipy>=1.9.1
soundfile>=0.11.0
torch>=1.12.1
torch-audiomentations==0.11.0
torchaudio>=0.12.1
tqdm>=4.64.1
